==> 2. Dependencies.Rmd <==
# Packages et librairies

```{r eval=FALSE}
# Installation alternative d'arrow pour MacOs
install.packages('arrow', repos = c('https://apache.r-universe.dev', 'https://cloud.r-project.org'))
```

```{r include=FALSE}
packages <- c(
  "archive",
  "arrow",
  "aws.s3",
  "cluster",
  "corrplot",
  "data.table",
  "factoextra",
  "FactoMineR",
  "ggcorrplot",
  "ggplot2",
  "ggridges",
  "gt",
  "httr",
  "quantreg",
  "magrittr",
  "mapview",
  "miceadds",
  "paletteer",
  "plm",
  "reticulate",
  "sf",
  "stargazer",
  "viridis",
  "xtable")

installed_packages <- packages %in% rownames(installed.packages())
if (any(installed_packages == FALSE)) {install.packages(packages[!installed_packages])}

invisible(lapply(packages, library, character.only = TRUE))

rm(installed_packages, packages)
```

# Répertoires

```{r}
if (!dir.exists("data_main")) {dir.create("data_main")}
if (!dir.exists("output_panel")) {dir.create("output_panel")}
if (!dir.exists("output_transversal")) {dir.create("output_transversal")}
if (!dir.exists("output_transversal/map_ppropri")) {dir.create("output_transversal/map_ppropri")}
if (!dir.exists("output_transversal/map_clusters")) {dir.create("output_transversal/map_clusters")}
```

# Fonctions

```{r}
# Pour charger les données Parquet
load_parquet <- function(filename) {
  local_filepath <- file.path("data_main/", filename)
  
  if (file.exists(local_filepath)) {
    message("Loading ", filename, " from local directory")
    data <- arrow::read_parquet(local_filepath)
    data <- as.data.table(data)
    return(data)
    
  } else {
    message("Downloading ", filename, " from SSP Cloud")
    tryCatch({
        s3_filepath <- file.path("/diffusion", filename)
        data <- aws.s3::s3read_using(
            FUN = arrow::read_parquet,
            object = s3_filepath,
            bucket = "maeldieudonne",
            opts = list("region" = ""))
        data <- as.data.table(data)
        return(data)
    }, error = function(e) {
      custom_message <- paste("Failed to download", filename, "from SSP cloud.")
      message(custom_message)
      stop(custom_message)})
  }
}
```

```{r}
# Pour charger les données Shapefile
load_shapefile <- function(filename) {
  local_filepath <- paste0("data_other/shapefile/", filename)
  
  if (file.exists(local_filepath)) {
    message("Loading from local directory")
    data <- sf::st_read(local_filepath)
    return(data)
    
  } else {
    message("Downloading from SSP cloud")
    if (!dir.exists("data_other/shapefile")) {dir.create("data_other/shapefile")}
    tryCatch({
      aws.s3::s3sync(
        path = "data_other/shapefile",
        bucket = "maeldieudonne",
        region = "",
        prefix = "diffusion/shapefile/",
        direction = c("upload", "download"),
        multipart = TRUE,
        create = TRUE)
      data <- sf::st_read(local_filepath)
      return(data)
    }, error = function(e) {
      custom_message <- paste("Failed to download", filename, "from SSP cloud.")
      message(custom_message)
      stop(custom_message)})
  }
}
```

```{r}
# Pour formater les nombres dans les tableaux
tidy_table <- function(x) {
  formatC(x, format = "f", digits = 3, drop0trailing = TRUE)
}
```

```{r}
# Pour générer des abréviations
get_first_letters <- function(input_string) {
  words <- unlist(strsplit(input_string, " |-"))
  if (length(words) == 1) {first_letters <- substr(words, 1, 3)} 
  else {first_letters <- substr(words, 1, 1)}
  paste(first_letters, collapse = "")
}
```

```{r}
# Pour générer des cartes statiques
# Cette fonction comporte 5 paramètres :
  # year
  # region permet de restreindre la carte à une région
    # La fonction utilise les codes listés dans data_other/code_regions_Insee.csv
    # La valeur 0 correspond à la France entière.
  # mapped_var est le code de la variable à représenter
  # mapped_var_name est le nom à afficher dans la légende
  # base_width est un paramètre de résolution, de préférence >7 pour une région et >15 pour la France
# Les cartes sont sauvegardées dans le répertoire output, leur nom est généré automatiquement.
# Elles sont aussi affichées ici dans le notebook.
# Pour la France entière, cela prend du temps !

draw_map <- function(input_year, region, mapped_var, mapped_var_name, base_width) {
  # Chargement des données
  carto <- subset(panel_data, year == input_year)
  carto <- merge(map_data, carto, by.x = "INSEE_COM", by.y = "codecommune", all.x = TRUE)
  if (region != 0) {
    carto <- subset(carto, INSEE_REG == region)
    reg_name <- codes_regions$LIBELLE[codes_regions$REG == region]
    reg_abrv <- codes_regions$ABRV[codes_regions$REG == region]
  } else {
    reg_name <- "France"
    reg_abrv <- "FR"
  }

  # Création des chemins
  title <- paste0(mapped_var_name, "\nPremier tour des élections présidentielles\n", reg_name, ", ", input_year)
  output_file <- paste0("output_transversal/", mapped_var, "_", input_year, "_", reg_abrv, ".png")
  
  # Exportation d'une carte en haute résolution
  ## Calcul des dimensions
  bbox <- st_bbox(carto)
  aspect_ratio <- diff(c(bbox$xmin, bbox$xmax)) / diff(c(bbox$ymin, bbox$ymax))
  base_height <- base_width / aspect_ratio
  font_size <- base_width * 2.2
  ## Création et enregistrement
  plot_png <- ggplot() +
    geom_sf(data = carto, aes(fill = .data[[mapped_var]])) +
    scale_fill_viridis_c(name = mapped_var_name, labels = scales::percent_format(scale = 1)) +
    ggtitle(paste(title)) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = 0.5, size = font_size),
          axis.text = element_blank(),
          panel.grid = element_blank(),
          plot.title.position = "plot",
          legend.title = element_text(size = font_size * 0.8),
          legend.text = element_text(size = font_size * 0.7))
  ggsave(output_file, plot = plot_png, width = base_width, height = base_height, units = "in")
  
  # Affichage d'une carte en résolution normale
  display_plot <- ggplot() +
    geom_sf(data = carto, aes(fill = .data[[mapped_var]])) +
    scale_fill_viridis_c(name = mapped_var_name, labels = scales::percent_format(scale = 1)) +
    ggtitle(paste(title)) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = 0.5),
          axis.text = element_blank(),
          panel.grid = element_blank(),
          plot.title.position = "plot",
          legend.title = element_text(),
          legend.text = element_text())   
    print(display_plot)
}
```
