==> 4. Fetch_data_other.Rmd <==
# Codes régionaux et départementaux

```{r}
httr::GET("https://www.insee.fr/fr/statistiques/fichier/6800675/v_region_2023.csv", timeout(15), progress(), write_disk("data_other/codes_regions_Insee.csv", overwrite = TRUE))
```

```{r}
httr::GET("https://www.insee.fr/fr/statistiques/fichier/6800675/v_departement_2023.csv", timeout(15), progress(), write_disk("data_other/codes_departements_Insee.csv", overwrite = TRUE))
```

# Données cartographiques

Téléchargement

```{r}
httr::GET("https://wxs.ign.fr/x02uy2aiwjo9bm8ce5plwqmr/telechargement/prepackage/ADMINEXPRESS_SHP_TERRITOIRES_PACK_2023-12-19$ADMIN-EXPRESS_3-2__SHP_LAMB93_FXX_2023-12-19/file/ADMIN-EXPRESS_3-2__SHP_LAMB93_FXX_2023-12-19.7z", timeout(1800), progress(), write_disk("data_other/Carto_Insee.7z", overwrite = TRUE))

# Extraction
root_name <- ("ADMIN-EXPRESS_3-2__SHP_LAMB93_FXX_2023-12-19/ADMIN-EXPRESS/1_DONNEES_LIVRAISON_2023-12-00156/ADE_3-2_SHP_LAMB93_FXX-ED2023-12-19/")
file_names <- c("COMMUNE", "ARRONDISSEMENT_MUNICIPAL")
extensions <- c("shp", "shx", "dbf")

files_to_extract <- c()
for (file_name in file_names) {
  for (ext in extensions) {
    files_to_extract <- c(files_to_extract, paste0(root_name, file_name, ".", ext))
  }
}

archive_extract("data_other/Carto_Insee.7z", dir = "data_other/shapefile", files = files_to_extract)

# Nettoyage
root_path <- file.path("data_other", "shapefile", "ADMIN-EXPRESS_3-2__SHP_LAMB93_FXX_2023-12-19", "ADMIN-EXPRESS", "1_DONNEES_LIVRAISON_2023-12-00156", "ADE_3-2_SHP_LAMB93_FXX-ED2023-12-19")

for (file_name in file_names) {
  for (ext in extensions) {
    source_file <- file.path(root_path, paste0(file_name, ".", ext))
    target_file <- file.path("data_other/shapefile", paste0(file_name, ".", ext))
    file.rename(source_file, target_file)
  }
}

unlink(c("data_other/Carto_Insee.7z", "data_other/shapefile/ADMIN-EXPRESS_3-2__SHP_LAMB93_FXX_2023-12-19"), recursive = TRUE)
rm(ext, extensions, file_name, file_names, files_to_extract, root_name, root_path, source_file, target_file)
```

Téléversement sur le Datalab

```{r eval=FALSE}
aws.s3::s3sync(
  path = "data_other/shapefile",
  bucket = "maeldieudonne",
  region = "",
  prefix = "/diffusion/shapefile/",
  direction = "upload",
  multipart = TRUE,
  create = TRUE)

# Nettoyage
unlink("data_other/shapefile", recursive = TRUE)
```
