==> 5. Prepare_data.Rmd <==
# Données de panel

L'enjeu est de construire une base unique à partir de données dispersées, entre des bases annuelles pour les résultats électoraux, proches du format long, et des bases globales pour les régresseurs, au format wide. La stratégie retenue consiste à itérer sur une table vide à l'aide des résultats électoraux.

On commence par créer une table de destination avec toutes les variables prévue, puis pour chaque base électorale :

1.  On recode les outcomes à l'aide des informations sur les orientations politiques des candidats, stockées dans un fichier dédié.

2.  On ajoute les régresseurs de l'année correspondante (renaming pour passer au format long puis merge).

3.  On fusionne avec la table de destination (rbind pour s'ajouter aux données existantes en éliminant les variables superflues).

L'inconvénient de cette méthode est qu'elle implique de charger simultanément toutes les bases, ce qui requiert beaucoup de mémoire. Elle simplifie toutefois la sélection des variables pertinentes, ainsi que le passage du long au format wide.

```{r}
# Chargement des données sur les orientations politiques des candidats
candidats <- fread("data_other/candidats.csv", header = TRUE)

# Chargement des régresseurs
propri <- load_parquet("Proprietaires/proprietairescommunes.parquet")
popul <- load_parquet("Taille_agglo_commune/popcommunes.parquet")
# pop_elec <- load_parquet("Taille_agglo_commune/popcommuneselecteurs.parquet")
ages <- load_parquet("Age_csp/agesexcommunes.parquet")
catsociopro <- load_parquet("CSP/cspcommunes.parquet")
nation <- load_parquet("Nationalites/etrangerscommunes.parquet")
reven <- load_parquet("Revenus/revcommunes.parquet")
diplo <- load_parquet("Diplomes/diplomescommunes.parquet")

# Conversion du nombre de (non-)diplômés en proportion du nombre d'habitants >25 ans
for (year in unique(candidats$Annee)) {
  suffixes <- c("propnodip", "propbac", "propsup", "nodip", "bac", "sup")
  for (i in seq_along(suffixes)) {assign(paste0("var_name_", i), paste0(suffixes[i], year))}
  diplo[, (var_name_1) := get(var_name_4)/(get(var_name_4) + get(var_name_5) + get(var_name_6))]
  diplo[, (var_name_2) := get(var_name_5)/(get(var_name_4) + get(var_name_5) + get(var_name_6))]
  diplo[, (var_name_3) := get(var_name_6)/(get(var_name_4) + get(var_name_5) + get(var_name_6))]
}
```

```{r}
# Création de la table de destination
panel_data <- data.table(
  dep = character(),
  codecommune = character(),
  year = integer(),
  # Outcomes
  part_T1 = numeric(),
  part_T2 = numeric(),
  GD_ratio_T1 = numeric(),
  GD_ratio_T2 = numeric(),
  GIns_ratio_T1 = numeric(),
  DIns_ratio_T1 = numeric(),
  GIns_ratio_T2 = numeric(),
  DIns_ratio_T2 = numeric(),
  EC_ratio_T1 = numeric(),
  EC_ratio_T2 = numeric(),
  EIns_ratio_T1 = numeric(),
  CIns_ratio_T1 = numeric(),
  EIns_ratio_T2 = numeric(),
  CIns_ratio_T2 = numeric(),  
  # Régresseurs
  ppropri = numeric(),
  popcomm = numeric(),
  popagglo = numeric (),
  age = numeric (),
  prop014 = numeric (),
  prop1539 = numeric (),
  prop4059 = numeric (),
  prop60p = numeric (),
  propf = numeric(),
  propnodip = numeric (),
  propbac = numeric (),
  propsup = numeric (),
  pagri = numeric (),
  pindp = numeric (),
  pcadr = numeric (),
  ppint = numeric (),
  pempl = numeric (),
  pouvr = numeric (),
  pchom = numeric (),
  petranger = numeric(),
  revratio = numeric()
  )
```

Nous avons renoncé à estimer le taux d'inscription sur les listes électorales. Il existe deux possibilités pour le faire : diviser le nombre d'inscrit par la population communale, ou par la population électorale estimée par les auteurs d'après les données socio-démographique. Mais chacune engendre de nombreuses valeurs supérieures à 1, souvent de beaucoup, ce qui les rend difficilement interprétable.

```{r}
# Fonction assurant la création des outcomes, l'ajout des régresseurs et la fusion avec la base de destination
recode_add_bind <- function(input_year) {
  tablename <- paste0("pres", input_year)
  dt <- get(tablename)
  # Ajout des années lorsqu'elles manquent dans la table d'origine
  if (year %in% colnames(dt)) {dt[is.na(dt$year), year := input_year]}
  else {dt[, year := input_year]}

  # Création des outcomes
  ## Participation
  dt[, part_T1 := votants/inscrits]
  dt[, part_T2 := votantsT2/inscritsT2]
  ## Droite / gauche
  ### Identification des candidats
  T1maskG <- candidats[Annee == input_year & DrGau == "G", .(variable = paste0("voix", Candidat))]
  T1maskD <- candidats[Annee == input_year & DrGau == "D", .(variable = paste0("voix", Candidat))]
  T2maskG <- candidats[Annee == input_year & DrGau == "G", .(variable = paste0("voixT2", Candidat))]
  T2candidateG <- intersect(T2maskG$variable, names(dt))
  T2maskD <- candidats[Annee == input_year & DrGau == "D", .(variable = paste0("voixT2", Candidat))]
  T2candidateD <- intersect(T2maskD$variable, names(dt))
  ### Ratio relatif / sur l'orientation politique opposée
  dt[, GD_ratio_T1 := rowSums(.SD, na.rm = TRUE) / exprimes, .SDcols = T1maskG$variable]
  dt[is.infinite(GD_ratio_T1), GD_ratio_T1 := 1]
  ### Pour le second tour, le ratio G/exp est seulement calculé si un candidat de gauche est présent.
  if (length(T2candidateG) == 1) {
    dt[, GD_ratio_T2 := rowSums(.SD, na.rm = TRUE) / exprimesT2, .SDcols = T2candidateG]
    dt[is.infinite(GD_ratio_T2), GD_ratio_T2 := 1]
    }
  else {dt[, GD_ratio_T2 := NA]}
  ### Ratios absolus / sur les inscrits
  dt[, GIns_ratio_T1 := rowSums(.SD, na.rm = TRUE) / inscrits, .SDcols = T1maskG$variable]
  dt[is.infinite(GIns_ratio_T1), GIns_ratio_T1 := 1]
  dt[, DIns_ratio_T1 := rowSums(.SD, na.rm = TRUE) / inscrits, .SDcols = T1maskD$variable]
  dt[is.infinite(DIns_ratio_T1), DIns_ratio_T1 := 1]
  ### Pour le second tour, le ratio de chaque orientation est seulement calculé si un candidat est présent. 
  if (length(T2candidateG) >= 1) {
    dt[, GIns_ratio_T2 := rowSums(.SD, na.rm = TRUE) / inscritsT2, .SDcols = T2candidateG]
    dt[is.infinite(GIns_ratio_T2), GIns_ratio_T2 := 1]
    } 
  else {dt[, GIns_ratio_T2 := NA]}
  if (length(T2candidateD) >= 1) {
    dt[, DIns_ratio_T2 := rowSums(.SD, na.rm = TRUE) / inscritsT2, .SDcols = T2candidateD]
    dt[is.infinite(DIns_ratio_T2), DIns_ratio_T2 := 1]
    } 
  else {dt[, DIns_ratio_T2 := NA]}
  ## Centre / extrêmes
  ### Identification des candidats
  T1maskC <- candidats[Annee == input_year & CentExt == "C", .(variable = paste0("voix", Candidat))]
  T1maskE <- candidats[Annee == input_year & CentExt == "E", .(variable = paste0("voix", Candidat))]
  T2maskC <- candidats[Annee == input_year & CentExt == "C", .(variable = paste0("voixT2", Candidat))]
  T2candidateC <- intersect(T2maskC$variable, names(dt))
  T2maskE <- candidats[Annee == input_year & CentExt == "E", .(variable = paste0("voixT2", Candidat))]
  T2candidateE <- intersect(T2maskE$variable, names(dt))
  ### Ratio relatif / sur l'orientation politique opposée
  dt[, EC_ratio_T1 := rowSums(.SD, na.rm = TRUE) / exprimes, .SDcols = T1maskE$variable]
  dt[is.infinite(EC_ratio_T1), EC_ratio_T1 := 1]
  ### Pour le second tour, le ratio E/exp est seulement calculé si un candidat des extrêmes est présent.
  if (length(T2candidateE) == 1) {
    dt[, EC_ratio_T2 := rowSums(.SD, na.rm = TRUE) / exprimesT2, .SDcols = T2candidateE]
    dt[is.infinite(EC_ratio_T2), EC_ratio_T2 := 1]
    }
  else {dt[, EC_ratio_T2 := NA]}
  ### Ratios absolus / sur les inscrits
  dt[, CIns_ratio_T1 := rowSums(.SD, na.rm = TRUE) / inscrits, .SDcols = T1maskC$variable]
  dt[is.infinite(CIns_ratio_T1), CIns_ratio_T1 := 1]
  dt[, EIns_ratio_T1 := rowSums(.SD, na.rm = TRUE) / inscrits, .SDcols = T1maskE$variable]
  dt[is.infinite(EIns_ratio_T1), EIns_ratio_T1 := 1]
  ### Pour le second tour, le ratio de chaque orientation est seulement calculé si un candidat est présent. 
  if (length(T2candidateC) >= 1) {
    dt[, CIns_ratio_T2 := rowSums(.SD, na.rm = TRUE) / inscritsT2, .SDcols = T2candidateC]
    dt[is.infinite(CIns_ratio_T2), CIns_ratio_T2 := 1]
    } 
  else {dt[, CIns_ratio_T2 := NA]}
  if (length(T2candidateE) >= 1) {
    dt[, EIns_ratio_T2 := rowSums(.SD, na.rm = TRUE) / inscritsT2, .SDcols = T2candidateE]
    dt[is.infinite(EIns_ratio_T2), EIns_ratio_T2 := 1]
    } 
  else {dt[, EIns_ratio_T2 := NA]}

  # Ajout des régresseurs
  ## Propriétaires
  var_name_1 <- paste0("ppropri", input_year)
  dt[propri, "ppropri" := .(get(var_name_1)), on = "codecommune"]
  ## Population
  var_name_1 <- paste0("pop", input_year)
  var_name_2 <- paste0("popagglo", input_year)
  dt[popul, c("popcomm", "popagglo") := .(get(var_name_1), get(var_name_2)), on = "codecommune"]
  ## Âge et sexe
  suffixes <- c("age", "prop014", "prop1539", "prop4059", "prop60p", "propf")
  for (suffix in suffixes) {
    var_name <- paste0(suffix, year)
    dt[ages, (suffix) := .(get(var_name)), on = "codecommune"]
  }
  ## Diplômes
  suffixes <- c("propnodip", "propbac", "propsup")
  for (suffix in suffixes) {
    var_name <- paste0(suffix, year)
    dt[diplo, (suffix) := .(get(var_name)), on = "codecommune"]
  }
  ## CSP
  suffixes <- c("pagri", "pindp", "pcadr", "ppint", "pempl", "pouvr", "pchom")
  for (suffix in suffixes) {
    var_name <- paste0(suffix, year)
    dt[catsociopro, (suffix) := .(get(var_name)), on = "codecommune"]
  }
  ## Étrangers
  var_name_1 <- paste0("petranger", input_year)
  dt[nation, "petranger" := .(get(var_name_1)), on = "codecommune"]
  ## Revenus communaux
  var_name_1 <- paste0("revratio", input_year)
  dt[reven, "revratio" := .(get(var_name_1)), on = "codecommune"]

  # Rbind
  cols_to_keep <- intersect(names(panel_data), names(dt))
  dt <- dt[, ..cols_to_keep]
  panel_data <- rbindlist(list(panel_data, dt))
  return(panel_data)
}
```

```{r}
# Boucle sur les élections
for (year in unique(candidats$Annee)) {
  output_file <- paste0("pres", year) 
  input_file <- paste0("Elections_pres/pres", year, "comm.parquet")
  assign(output_file, load_parquet(input_file))
  panel_data <- recode_add_bind(year)
  cat("Data generated for year", year, "\n")
}
```

```{r}
# Étonnamment une trentaine de valeurs >1 apparaissent pour la proportion de chômeurs.
# Elles ne semblent pas présentes dans la table initiale, dont les valeurs ne font l'objet d'aucune transformation.
# Faute d'explication, et devant leur faible nombre, on les remplace par des NA.
panel_data[pchom > 1, pchom := NA]

# Une entrée dépourvue de valeur pour le codecommune apparaît aussi : on la supprime.
panel_data <- panel_data[nchar(panel_data$codecommune) >= 5, ]

# On élimine les arrondissements de Lyon et Marseille.
panel_data <- subset(panel_data, !(codecommune %in% c(69381:69389, 13201:13216)))
```

```{r}
# Conversion des proportions en pourcentages pour faciliter la lecture des résultats
excluded_variables <- c("dep", "codecommune", "year", "popcomm", "popagglo", "age", "revratio")
variables_to_convert <- setdiff(names(panel_data), excluded_variables)
panel_data[, (variables_to_convert) := lapply(.SD, function(x) x * 100), .SDcols = variables_to_convert]

# Expression de la population et des revenus en milliers
variables_to_convert <- c("popcomm", "popagglo")
panel_data[, (variables_to_convert) := lapply(.SD, function(x) x / 1000), .SDcols = variables_to_convert]

# Ecriture du fichier
write_parquet(panel_data, "data_main/panel_data.parquet")
```

```{r}
# Arrière-pensée : construction d'une base additionnelle au format wide pour les tours
# Donc où l'on n'a plus d'outcomes pour chaque tour, mais des outcomes globaux et une indicatrice du tour (égale à 1 pour le second tour)
colnames <- colnames(panel_data)
cols_to_keep <- colnames[!grepl("T1$|T2$", colnames)]

regressors_T1 <- panel_data[, ..cols_to_keep, drop = FALSE]
regressors_T1$round <- 1
regressors_T1$round <- as.character(regressors_T1$round)
regressors_T2 <- panel_data[, ..cols_to_keep, drop = FALSE]
regressors_T2$round <- 2
regressors_T2$round <- as.character(regressors_T2$round)
wide_panel_data <- rbind(regressors_T1, regressors_T2)

part_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("part_T1", "part_T2"), variable.name = "round", value.name = "part")
part_panel_data$round <- ifelse(!is.na(part_panel_data$round), sub("^part_T", "", part_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, part_panel_data, by = c("codecommune", "year", "round"))

GD_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("GD_ratio_T1", "GD_ratio_T2"), variable.name = "round", value.name = "GD_ratio")
GD_panel_data$round <- ifelse(!is.na(GD_panel_data$round), sub("^GD_ratio_T", "", GD_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, GD_panel_data, by = c("codecommune", "year", "round"))

EC_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("EC_ratio_T1", "EC_ratio_T2"), variable.name = "round", value.name = "EC_ratio")
EC_panel_data$round <- ifelse(!is.na(EC_panel_data$round), sub("^EC_ratio_T", "", EC_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, EC_panel_data, by = c("codecommune", "year", "round"))

GIns_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("GIns_ratio_T1", "GIns_ratio_T2"), variable.name = "round", value.name = "GIns_ratio")
GIns_panel_data$round <- ifelse(!is.na(GIns_panel_data$round), sub("^GIns_ratio_T", "", GIns_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, GIns_panel_data, by = c("codecommune", "year", "round"))

DIns_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("DIns_ratio_T1", "DIns_ratio_T2"), variable.name = "round", value.name = "DIns_ratio")
DIns_panel_data$round <- ifelse(!is.na(DIns_panel_data$round), sub("^DIns_ratio_T", "", DIns_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, DIns_panel_data, by = c("codecommune", "year", "round"))

CIns_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("CIns_ratio_T1", "CIns_ratio_T2"), variable.name = "round", value.name = "CIns_ratio")
CIns_panel_data$round <- ifelse(!is.na(CIns_panel_data$round), sub("^CIns_ratio_T", "", CIns_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, CIns_panel_data, by = c("codecommune", "year", "round"))

EIns_panel_data = melt(panel_data, id.vars = c("codecommune", "year"), measure.vars = c("EIns_ratio_T1", "EIns_ratio_T2"), variable.name = "round", value.name = "EIns_ratio")
EIns_panel_data$round <- ifelse(!is.na(EIns_panel_data$round), sub("^EIns_ratio_T", "", EIns_panel_data$round), NA)
wide_panel_data <- merge(wide_panel_data, EIns_panel_data, by = c("codecommune", "year", "round"))

wide_panel_data[, round := ifelse(round == 1, 0, ifelse(round == 2, 1, round))]
wide_panel_data$round <- as.numeric(wide_panel_data$round)

write_parquet(wide_panel_data, "data_main/wide_panel_data.parquet")
```

```{r}
# Nettoyage
## Suppression des résultats électoraux
for (year in unique(candidats$Annee)) {
  tablename <- paste0("pres", year)
  rm(list = tablename)}

## Suppression des données de contrôle
rm(ages, candidats, catsociopro, diplo, nation, popul, propri, reven)

## Suppression des variables
for (name in paste0("var_name_", 1:6)) {rm(list = name)}
rm(i, input_file, name, tablename, output_file, suffixes, year)
rm(excluded_variables, variables_to_convert)

## Supression des tables intermédiaires dans la séparation des tours
rm(colnames, cols_to_keep, regressors_T1, regressors_T2, part_panel_data, GD_panel_data, EC_panel_data, GIns_panel_data, DIns_panel_data, CIns_panel_data, EIns_panel_data)
```

```{r eval=FALSE}
# Téléversement sur le SSP Cloud
aws.s3::put_object(
  "data_main/panel_data.parquet",
  object = "/diffusion/panel_data.parquet",
  bucket = "maeldieudonne",
  region = "",
  multipart = TRUE,
  overwrite = TRUE)

aws.s3::put_object(
  "data_main/wide_panel_data.parquet",
  object = "/diffusion/wide_panel_data.parquet",
  bucket = "maeldieudonne",
  region = "",
  multipart = TRUE,
  overwrite = TRUE)
```

# Données de coupe

On construit également une base restreinte à une année, mais intégrant toutes les données disponibles seulement en fin de période : sur la criminalité, la fiscalité, les prestations sociales, ainsi que les types de communes. On part de la base précédente en sélectionnant l'année 2017, puis l'on ajoute les régresseurs supplémentaires.

```{r}
# Chargement des bases
crim <- load_parquet("CSP/crimesdelitscommunes.parquet")
isf <- load_parquet("Capital_immobilier/isfcommunes.parquet")
reven <- load_parquet("Revenus/revcommunes.parquet")
rsa <- load_parquet("CSP/rsacommunes.parquet")
vbbm <- load_parquet("Taille_agglo_commune/popcommunesvbbm.parquet")

# Correction des codes communes dans la base RSA (traités comme des nombres donc avec suppression des zéros initiaux)
rsa$codecommune <- as.character(rsa$codecommune)
rsa$codecommune <- ifelse(nchar(rsa$dep) == 1, paste0("0", rsa$codecommune), rsa$codecommune)

# Sélection de l'année 2017 dans les données de panel et ajout de régresseurs supplémentaires
mca_data <- subset(panel_data, year == 2017)
mca_data[crim, "pcrimesdelits" := i.pcrimesdelits2016, on = "codecommune"]
mca_data[isf, c("mmoyfortune", "pisf") := .(mmoyfortune2017, pisf2017), on = "codecommune"]
mca_data[rsa, "prsa" := i.prsa2017, on = "codecommune"]
mca_data[vbbm, "type_comm" := i.vbbm2017, on = "codecommune"]

# Suppression du revenu commmunal estimé, remplacé par sa valeur réelle
mca_data <- mca_data[, revratio := NULL]
mca_data[reven, "revmoy" := i.revmoy2017, on = "codecommune"]

## Les données fiscales sont disponibles uniquement pour les communes de plus de 20 000 habitants comptant au moins 50 contribuables assujettis à l'ISF (soit 374 communes rassemblant 22,3 millions d’habitants en 2017).
## Les valeurs manquantes sont remplacées par des valeurs nulles.
mca_data[is.na(mmoyfortune), mmoyfortune := 0]
mca_data[is.na(pisf), pisf := 0]

# Conversion des ratios en pourcentages
variables_to_convert <- c("pisf", "prsa")
mca_data[, (variables_to_convert) := lapply(.SD, function(x) x * 100), .SDcols = variables_to_convert]

# Expression de la fortune moyenne des contribuables soumis à l'ISF en milliers
mca_data[, mmoyfortune := mmoyfortune / 1000]

# Ecriture du fichier et nettoyage
write_parquet(mca_data, "data_main/mca_data.parquet")
rm(crim, isf, reven, rsa, vbbm)
```

```{r}
# Ajout éventuel : les indicateurs de religiosité (qui ont l'inconvénient d'être anciens et d'avoir beaucoup de valeurs manquantes)

# Chargement des bases
religcomm <- load_parquet("Enseignement_prive/religiositecommunes.parquet")
religdep <- load_parquet("Enseignement_prive/religiositedepartements.parquet")
popul <- load_parquet("Taille_agglo_commune/popcommunes.parquet")

# Ratio de nclerge1791 sur le nombre d'habitants, pour avoir une idée de la présence des institutions religieuses dans la commune
religcomm[popul, c("pop1791") := .(i.pop1791), on = "codecommune"]
religcomm[, pclerge1791 := nclerge1791/pop1791]
religcomm[, pclerge1791 := replace(pclerge1791, is.infinite(pclerge1791), NA)]

# Ajout aux données
mca_data[religcomm, c("prefract1791", "pclerge1791") :=
             .(i.prefract1791, i.pclerge1791), on = "codecommune"]
mca_data[religdep, c("pclerge1856", "pmessalisants1950") :=
             .(i.pclerge1856, i.pmessalisants1950), on = "dep"]

# Conversion des ratios en pourcentages
variables_to_convert <- c("prefract1791", "pclerge1791", "pclerge1856", "pmessalisants1950")
mca_data[, (variables_to_convert) := lapply(.SD, function(x) x * 100), .SDcols = variables_to_convert]

# Ecriture du fichier et nettoyage
write_parquet(mca_data, "data_main/mca_data.parquet")
rm(popul, religcomm, religdep, variables_to_convert)
```

```{r}
# Seconde arrière-pensée : construction d'une base additionnelle avec les résultats électoraux détaillés plutôt qu'agrégés
pres_2017 <- load_parquet("Elections_pres/pres2017comm.parquet")
cols_to_keep <- grep("^pvoix.*[^ratio]$", names(pres_2017), value = TRUE)
cols_to_keep <- c("codecommune", cols_to_keep)
pres_2017 <- pres_2017[, ..cols_to_keep, with = FALSE]
pres_2017[, (grep("^pvoix", names(pres_2017), value = TRUE)) := lapply(.SD, function(x) x * 100), .SDcols = grep("^pvoix", names(pres_2017), value = TRUE)]

cols_to_drop <- c("GD_ratio_T1", "GD_ratio_T2", "EC_ratio_T1", "EC_ratio_T2", "GIns_ratio_T1", "DIns_ratio_T1", "GIns_ratio_T2", "DIns_ratio_T2", "EIns_ratio_T1", "CIns_ratio_T1", "EIns_ratio_T2", "CIns_ratio_T2")
cols_to_keep <- setdiff(names(mca_data), cols_to_drop)
detailed_mca_data <- mca_data[, ..cols_to_keep, with = FALSE]

detailed_mca_data <- merge(detailed_mca_data, pres_2017, by = "codecommune")

write_parquet(detailed_mca_data, "data_main/detailed_mca_data.parquet")

rm(cols_to_drop, cols_to_keep, pres_2017)
```

```{r eval=FALSE}
# Téléversement sur le SSP Cloud
aws.s3::put_object(
  "data_main/mca_data.parquet",
  object = "/diffusion/mca_data.parquet",
  bucket = "maeldieudonne",
  region = "",
  multipart = TRUE,
  overwrite = TRUE)

aws.s3::put_object(
  "data_main/detailed_mca_data.parquet",
  object = "/diffusion/detailed_mca_data.parquet",
  bucket = "maeldieudonne",
  region = "",
  multipart = TRUE,
  overwrite = TRUE)
```

# Cartes

Les données géographiques sont très volumineuses car très précises, et stockées dans un format spécifique. Il serait donc malcommode des les intégrer aux bases précédentes. On prévoit plutôt d'utiliser une fonction construisant dynamiquement la base nécessaire chaque fois qu'une carte est tracée.

On adapte les données géographiques à cette fin. Il faut assembler celles relatives aux communes et aux arrondissements, car les données électorales sont présentées à l'échelle des arrondissements pour Paris, et des communes entières pour Lyon et Marseille. On ajoute aussi les codes régionaux de l'Insee afin de pouvoir tracer les cartes à l'échelle régionale, qui s'avère la plus lisible.

```{r}
background_province <- load_shapefile("COMMUNE.shx")
background_province <- subset(background_province, select = -c(INSEE_ARR, INSEE_CAN, INSEE_DEP, SIREN_EPCI, STATUT))

background_plm <- load_shapefile("ARRONDISSEMENT_MUNICIPAL.shx")
background_plm$INSEE_REG <- ifelse(startsWith(background_plm$INSEE_ARM, "13"), 93,
                                    ifelse(startsWith(background_plm$INSEE_ARM, "69"), 84,
                                           ifelse(startsWith(background_plm$INSEE_ARM, "75"), 11, NA)))
background_plm <- subset(background_plm, INSEE_REG == 11)
background_plm <- subset(background_plm, select = -INSEE_COM)
setnames(background_plm, old = "INSEE_ARM", new = "INSEE_COM")

map_data <- rbind(background_province, background_plm)
map_data <- subset(map_data, select = -c(ID, NOM, NOM_M, POPULATION))
st_crs(map_data) <- 9794

write_sf(map_data, "data_other/shapefile/FRANCE.shp")

rm(background_province, background_plm)
```

```{r eval=FALSE}
# Téléversement sur le SSP Cloud
aws.s3::s3sync(
  path = "data_other/shapefile",
  bucket = "maeldieudonne",
  region = "",
  prefix = "/diffusion/shapefile/",
  direction = "upload",
  multipart = TRUE,
  create = TRUE)
```
