==> 7. Panel_analysis.Rmd <==
```{r include=FALSE}
# Importation des données
panel_data <- load_parquet("panel_data.parquet")
wide_panel_data <- load_parquet("wide_panel_data.parquet")
map_data <- load_shapefile("FRANCE.shx")

codes_regions <- fread("data_other/codes_regions_Insee.csv", header = TRUE)
codes_regions$ABRV <- sapply(codes_regions$LIBELLE, get_first_letters)

# Création d'un identifiant année-tour pour analyser les données wide en panel
wide_panel_data$year_round <- with(wide_panel_data, paste(year, round, sep="-"))
```

```{r}
# Définition des variables
## Outcomes
relative_outcomes <- c("part_T1", "GD_ratio_T1", "EC_ratio_T1")
relative_outcomes_2T <- c("part_T1", "part_T2", "GD_ratio_T1", "GD_ratio_T2", "EC_ratio_T1", "EC_ratio_T2")
absolute_outcomes <- c("GIns_ratio_T1", "DIns_ratio_T1", "EIns_ratio_T1", "CIns_ratio_T1")
absolute_outcomes_2T <- c("GIns_ratio_T1", "DIns_ratio_T1", "GIns_ratio_T2", "DIns_ratio_T2", "EIns_ratio_T1", "CIns_ratio_T1", "EIns_ratio_T2", "CIns_ratio_T2")
wide_relative_outcomes <- c("part", "GD_ratio", "EC_ratio")

## Régresseurs
## Modalités supprimées pour éviter la colinéarité : prop014, propnodip, pouvr
panel_regressors <- c("ppropri", "popcomm", "popagglo", "age", "prop1539", "prop4059", "prop60p", "propf", "propbac", "propsup", "pagri", "pindp", "pempl", "ppint", "pcadr", "pchom", "petranger", "revratio")
full_panel_regressors <- c("ppropri", "popcomm", "popagglo", "age", "prop014", "prop1539", "prop4059", "prop60p", "propf", "propnodip", "propbac", "propsup", "pagri", "pindp", "pouvr", "pempl", "ppint", "pcadr", "pchom", "petranger", "revratio")
wide_panel_regressors <- c("round", panel_regressors)
```

# Analyse descriptives

## Valeurs manquantes

On vérifie facilement l'absence de valeurs manquantes pour les variables identifiantes (année et code commune).

```{r}
print(sum(is.na(panel_data$codecommune)))
print(sum(is.na(panel_data$year)))
```

Tableaux présentant la proportion de valeurs manquantes par outcome relatif, puis par régresseur, et par an.

```{r}
count_na <- function(x) mean(is.na(x)) * 100

variable_names = c(relative_outcomes_2T, full_panel_regressors)

na_prop_table <- data.table(year = unique(panel_data$year))
for (name in variable_names) {
  na_prop <- panel_data[, .(na_prop = count_na(get(name))), by = year]
  na_prop_table <- merge(na_prop_table, na_prop, by = "year", all.x = TRUE)
  setnames(na_prop_table, "na_prop", name)
}

na_prop_table <- na_prop_table %>%
  replace(., . == 100, NaN) %>%
  t() %>%
  `colnames<-`(.[1, ]) %>%
  .[-1, ]

max_value <- max(na_prop_table[is.finite(na_prop_table)])
bins <- seq(0, max_value, by = max_value/9)

gt_na_prop <- na_prop_table %>%
  as.data.frame() %>%
  gt(., rownames_to_stub = TRUE) %>%
  fmt_number(decimals = 2, drop_trailing_zeros = TRUE) %>%
  tab_header(title = "Proportion de valeurs manquantes", subtitle = NULL, preheader = NULL) %>%
  data_color(method = "bin", bins = bins, direction = "column", palette ="Oranges", na_color = "white", alpha = 0.6) %>%
  tab_row_group(label = "Outcomes", rows = 1:6, id = "out") %>%
  tab_row_group(label = "Régresseurs", rows = 7:27, id = "reg") %>%
  row_group_order(groups = c("out", "reg")) %>%
  sub_missing(missing_text = "-") %>%
  tab_style(
    style = list(
      cell_fill(color = "lightgrey"),
      cell_text(style = "italic", align = "center")),
    locations = cells_row_groups(groups = c("out", "reg"))) %>%
  tab_style(
    style = cell_text(align = "center"),
    locations = cells_body())

gt_na_prop
gtsave(gt_na_prop, filename = "nan_table.tex", path = "output_panel")

rm(count_na, variable_names, na_prop_table, max_value, bins, gt_na_prop, name, na_prop)
```

## Outcomes

```{r}
graph_data <- data.frame(year = unique(panel_data$year))

for (outcome in relative_outcomes_2T) {
  data <- aggregate(as.formula(paste(outcome, " ~ year")), data = panel_data, FUN = mean)
  graph_data <- merge(graph_data, data, by = "year", all = TRUE)
}

graph_data$year <- as.numeric(as.character(graph_data$year))

plot <- ggplot(graph_data, aes(x = year)) +
  geom_line(aes(y = part_T1, color = "part_T1"), na.rm = TRUE) +
  geom_line(aes(y = part_T2, color = "part_T2"), na.rm = TRUE) +
  geom_line(aes(y = GD_ratio_T1, color = "GD_ratio_T1"), na.rm = TRUE) +
  geom_line(aes(y = GD_ratio_T2, color = "GD_ratio_T2"), na.rm = TRUE) +
  geom_line(aes(y = EC_ratio_T1, color = "EC_ratio_T1"), na.rm = TRUE) +
  geom_line(aes(y = EC_ratio_T2, color = "EC_ratio_T2"), na.rm = TRUE) +
  scale_x_continuous(breaks = unique(graph_data$year)) +
  labs(
    x = "Années",
    y = "Pourcentage",
    color = "Outcomes") +
  scale_color_manual(values = c("part_T1" = "blue",
                                "part_T2" = "skyblue",
                                "GD_ratio_T1" = "limegreen",
                                "GD_ratio_T2" = "lightgreen",
                                "EC_ratio_T1" = "orange",
                                "EC_ratio_T2" = "red"),
                     labels = c("part_T1" = "Participation T1",
                                "part_T2" = "Participation T2",
                                "GD_ratio_T1" = "Ratio GD T1",
                                "GD_ratio_T2" = "Ratio GD T2",
                                "EC_ratio_T1" = "Ratio EC T1",
                                "EC_ratio_T2" = "Ratio EC T2")) +
  theme_grey() +
  theme(
    legend.title = element_blank(),
    panel.grid.minor = element_blank()
  )
ggsave("output_panel/outcomes_relatifs_1965_2022.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(
    title = "Evolution des paramètres d'intérêt",
    subtitle = "1965-2022") +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5))
print(plot)


rm(data, graph_data, outcome)
```

## Taux de propriétaires

Nous pouvons d'abord représenter l'évolution du taux national de propriétaires, calculé sur l'ensemble de la population française. Il n'est cependant pas le plus pertinent, car nous allons raisonner à l'échelle des communes plutôt que des individus. Représentons donc aussi l'évolution de la proportion moyenne de propriétaires dans chaque commune.

```{r}
propri <- load_parquet("Proprietaires/proprietairescommunes.parquet")
propri_df <- data.table(year = numeric(), sum_npropri = numeric(), sum_nlogement = numeric(), prop_ratio = numeric(), ppropri = numeric())

for (year in 1960:2022) {
  col_npropri <- paste0("npropri", year)
  col_nlogement <- paste0("nlogement", year)
  col_ppropri <- paste0("ppropri", year)
  sum_npropri <- sum(propri[[col_npropri]], na.rm = TRUE)
  sum_nlogement <- sum(propri[[col_nlogement]], na.rm = TRUE)
  average_ppropri <- 100 * mean(propri[[col_ppropri]], na.rm = TRUE)
  prop_ratio <- 100 * sum_npropri / sum_nlogement
  propri_df <- rbind(propri_df, data.table(year = year, sum_npropri = sum_npropri, sum_nlogement = sum_nlogement, prop_ratio = prop_ratio, ppropri = average_ppropri))
}

plot <- ggplot(propri_df, aes(x = year)) +
  geom_line(aes(y = prop_ratio, color = "prop_ratio")) +
  geom_line(aes(y = ppropri, color = "ppropri")) +
  labs(x = "Années", y = "", color = "") +
  scale_color_manual(
    values = c("prop_ratio" = "skyblue", "ppropri" = "orange"),
    labels = c("prop_ratio" = "Taux national", "ppropri" = "Taux communal moyen")) +
  theme_grey()
ggsave("output_panel/taux_proprietaires_1960_2022.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(
    title = "Taux de propriétaires de leur résidence principale",
    subtitle = "1960-2022") +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5))
print(plot)

rm(propri, propri_df, year, col_npropri, col_nlogement, col_ppropri, sum_npropri, sum_nlogement, average_ppropri, prop_ratio)
```

La différence entre les deux courbes atteste que les propriétaires de leur résidence principale sont proportionnellement plus nombreux dans les communes de faible taille. Pour caractériser davantage cette situation, représentons l'évolution des déciles du taux communal de propriétaires. On observe une certaine convergence, avec un ratio D1/D10 qui passe de 0,5 à 0,66 sur la période.

```{r}
nb_quantiles <- 11
quantiles_list <- list()

for (yr in unique(panel_data$year)) {
  panel_subset <- panel_data[panel_data$year == yr, ]
  quantiles <- quantile(panel_subset$ppropri, probs = seq(0, 1, length.out = nb_quantiles), na.rm = TRUE)
  names(quantiles) <- paste0("D", 0:(nb_quantiles - 1))
  quantiles_list[[as.character(yr)]] <- quantiles
}

quantiles_df <- do.call(rbind, quantiles_list)
quantiles_df <- data.table(year = as.integer(row.names(quantiles_df)), quantiles_df)
quantiles_df <- melt(quantiles_df, id.vars = "year", variable.name = "quintile")
names(quantiles_df) <- c("year", "quintile", "ppropri")
quantiles_df <- quantiles_df[quantiles_df$quintile != "D0" & quantiles_df$quintile != paste0("D", nb_quantiles - 1), ]

plot <- ggplot(quantiles_df, aes(x = year, y = ppropri, color = as.factor(quintile), group = quintile)) +
  geom_line() +
  labs(x = "Années",y = "", color = "Déciles") +
  theme_grey()

ggsave("output_panel/deciles_taux_proprietaires_1960_2022.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(title = "Déciles du taux communal moyen de propriétaires") +
  theme(plot.title = element_text(hjust = 0.5))
print(plot)


rm(nb_quantiles, quantiles_list, yr, panel_subset, quantiles, quantiles_df)
```

Pour terminer, un ridgeline représentant plus finement l'évolution de la distribution du taux de propriétaires.

```{r}
plot <- ggplot(panel_data, aes(x = ppropri, y = as.factor(year), fill = after_stat(x))) +
  geom_density_ridges_gradient(scale = 4, rel_min_height = 0.01) +
  scale_fill_viridis(option = "C") +
  labs(x = "", y = "") +
  theme_grey() +
  theme(legend.position = "none")
ggsave("output_panel/ridge_ppropri_1965_2022.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(
    title = "Distribution du taux de propriétaires",
    subtitle = "1965 - 2022") +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5))
print(plot)

rm(plot)
```

Étudions la corrélation du taux de propriétaires avec les autres régresseurs. Certains résultats sont peu surprenants : les propriétaires sont comparativement plus nombreux dans les communes dont les habitants sont plus âgés, et moins nombreux dans les communes plus petites et dont les habitants comptent davantage d'étrangers. Les corrélations avec le diplôme et la catégorie socio-professionnelles sont plus ambigues, avec une corrélation plus forte avec la proportion de diplômés du bac que du supérieur, et de professions intermédiaires que de cadres. On peut supposer que les diplômés du supérieur et les cadres résident davantage dans les grandes villes, ce qui freine leur accession à la propriété malgré leurs revenus supérieurs.

```{r}
panel_data_complete <- na.omit(panel_data[, c("ppropri", full_panel_regressors), with = FALSE])
correlation_results <- sapply(panel_data_complete[, -1, drop = FALSE], function(x) cor(panel_data_complete$ppropri, x))
correlation_results <- correlation_results * 100
correlation_df <- data.frame(Variables = names(correlation_results), Corrélation = correlation_results)
correlation_df <- subset(correlation_df, Variables != "ppropri")
correlation_df <- correlation_df[order(correlation_df$Corrélation, decreasing = TRUE), ]
rownames(correlation_df) <- NULL

correlation_df <- correlation_df %>%
  gt() %>%
  fmt_number(decimals = 2, drop_trailing_zeros = TRUE) %>%
  tab_header("Corrélation des régresseurs avec le taux de propriétaires", subtitle = NULL, preheader = NULL) %>%
  data_color(method = "numeric", columns = !matches("Variables"), palette ="Spectral", alpha = 0.6) %>%
  tab_style(
    style = cell_text(align = "center"),
    locations = list(
      cells_body(column = c("Corrélation")),
      cells_column_labels()))

correlation_df
gtsave(correlation_df, filename = "correl_ppropri.tex", path = "output_panel")

rm(panel_data_complete, correlation_results, correlation_df)
```

# Panel

## Tests

On utilise deux tests :

-   Le test de Wooldrige à partir des différences premières, qui concerne la corrélation temporelles des résidus.

-   Le test de Hausman, qui permet de comparer les estimateurs fd et within, donc d'évaluer l'hypothèse d'exogénéité stricte.

```{r}
# Hausman test
# Comparaison des estimateurs fd et within => H0 correspond à l'exogénéité stricte
panel_as_pdata <- pdata.frame(panel_data, index = c("codecommune", "year"))

print("Test for part_T1")
regression_formula <- as.formula(paste("part_T1 ~", paste(panel_regressors, collapse = "+")))
pwfdtest(regression_formula, data = panel_as_pdata)
fd <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
within <- plm(regression_formula, data = panel_as_pdata, model = "within", effect = "individual", cluster = "dep", vcov = "HC1")
phtest(fd, within)

print("Test for GD_ratio_T1")
regression_formula <- as.formula(paste("GD_ratio_T1 ~", paste(panel_regressors, collapse = "+")))
pwfdtest(regression_formula, data = panel_as_pdata)
fd <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
within <- plm(regression_formula, data = panel_as_pdata, model = "within", effect = "individual", cluster = "dep", vcov = "HC1")
phtest(fd, within)

print("Test for EC_ratio_T1")
regression_formula <- as.formula(paste("EC_ratio_T1 ~", paste(panel_regressors, collapse = "+")))
pwfdtest(regression_formula, data = panel_as_pdata)
fd <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
within <- plm(regression_formula, data = panel_as_pdata, model = "within", effect = "individual", cluster = "dep", vcov = "HC1")
phtest(fd, within)

rm(panel_as_pdata, fd, within)
```

## T1 seul

### Panels simples

On commence par des panels simples, en différences premières, avec clustering au niveau départemental et variance hétéroscédastique.

```{r}
# Pour les outcomes relatifs = ratios / exprimés
panel_as_pdata <- pdata.frame(panel_data, index = c("codecommune", "year"))

relative_results <- list()
for (outcome in relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, outcome, panel_as_pdata)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_panel_basic.tex"))
rm(relative_results)
```

```{r eval=FALSE}
# Pour les outcomes absolus = ratios / inscrits
panel_as_pdata <- pdata.frame(panel_data, index = c("codecommune", "year"))

absolute_results <- list()
for (outcome in absolute_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  absolute_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep")
}

suppressWarnings(stargazer(absolute_results, type = "text"))
rm(model_name, regression_formula, outcome, panel_as_pdata)
```

```{r eval=FALSE, include=FALSE}
suppressWarnings(stargazer(absolute_results, type = "latex", out = "output_panel/absolute_panel_basic.tex"))
rm(absolute_results)
```

On obtient des R2 très faibles, entre 0,5 et 2,8 %, et les coefficients de ppropri, bien que significatifs, ne sont pas toujours du signe attendu. Pour remédier à ces problèmes, on essaye d'introduire des dummies annuelles.

### Avec dummies temporelles

```{r}
# Pour les outcomes relatifs = ratios / exprimés
panel_as_pdata <- pdata.frame(panel_data, index = c("codecommune", "year"))
panel_as_pdata$year <- as.factor(panel_as_pdata$year)
regressors <- c(panel_regressors, "year")

relative_results <- list()
for (outcome in relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, regressors, outcome, panel_as_pdata)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_panel_timedummies.tex"))
rm(relative_results)
```

```{r eval=FALSE}
# Pour les outcomes absolus = ratios / inscrits
panel_as_pdata <- pdata.frame(panel_data, index = c("codecommune", "year"))
panel_as_pdata$year <- as.factor(panel_as_pdata$year)
regressors <- c(panel_regressors, "year")

absolute_results <- list()
for (outcome in absolute_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  absolute_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(absolute_results, type = "text"))
rm(model_name, regression_formula, regressors, outcome, panel_as_pdata)
```

```{r eval=FALSE, include=FALSE}
suppressWarnings(stargazer(absolute_results, type = "latex", out = "output_panel/absolute_panel_timedummies.tex"))
rm(absolute_results)
```

On obtient une très forte amélioration du R2, entre 21,6 et 63,9 %. Les effets de ppropri cessent d'être significatifs sur la participation. Ils le restent sur les autres outcomes, mais n'acquièrent pas de signification évidente (effet positif sur le vote à gauche, positif sur le vote au centre au T1 et négatif au T2). On teste l'hypothèse que les effets de ppropri évoluent au cours du temps en introduisant des interactions avec les dummies annuelles.

### Avec dummies temporelles et interactions

```{r}
# Pour les outcomes relatifs = ratios / exprimés
time_dummies <- model.matrix(~ year * ppropri - 1, data = panel_data)
panel_as_pdata <- pdata.frame(cbind(panel_data, time_dummies), index = c("codecommune", "year"))
panel_as_pdata$year <- as.factor(panel_as_pdata$year)

regressors <- c(panel_regressors, colnames(time_dummies))
regressors <- regressors[regressors != "ppropri"]

relative_results <- list()
for (outcome in relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, regressors, time_dummies, outcome, panel_as_pdata)
```

```{r}
get_coeffs <- function(model, outcome_name) {
  coefficients <- as.data.frame(model$coefficients)
  coefficients <- coefficients[grep("ppropri$", rownames(coefficients)), , drop = FALSE]
  colnames(coefficients) <- "coef_ppropri"
  years <- sub("^year(\\d+):.*", "\\1", rownames(coefficients))
  coefficients$year <- years
  coefficients$outcome <- outcome_name
  
  variance <- as.data.frame(model$vcov)
  variance <- variance[grep("ppropri$", rownames(variance)), grep("ppropri$", colnames(variance)), drop = FALSE]
  variance <- diag(as.matrix(variance))
  variance <- as.data.frame(variance)
  colnames(variance) <- "var"
  
  coefficients <- cbind(coefficients, variance)
  coefficients$CI_low <- coefficients$coef_ppropri - 1.96 * sqrt(coefficients$var)
  coefficients$CI_up <- coefficients$coef_ppropri + 1.96 * sqrt(coefficients$var)
  
  return(coefficients)
}

part_T1 <- get_coeffs(relative_results$part_T1_model, "part_T1")
GD_ratio_T1 <- get_coeffs(relative_results$GD_ratio_T1_model, "GD_ratio_T1")
EC_ratio_T1 <- get_coeffs(relative_results$EC_ratio_T1_model, "EC_ratio_T1")
coefficients <- rbind(part_T1, GD_ratio_T1, EC_ratio_T1)
coefficients$year <- as.numeric(coefficients$year)

plot <- ggplot(coefficients) +
  geom_line(aes(x = year, y = coef_ppropri, group = outcome, color = outcome)) +
  geom_line(aes(x = year, y = CI_low, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) +
  geom_line(aes(x = year, y = CI_up, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) + 
  geom_ribbon(aes(x = year, ymin = CI_low, ymax = CI_up, group = outcome, fill = outcome), alpha = 0.3) +
  geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
  scale_color_manual(
    values = c("part_T1" = "blue", "GD_ratio_T1" = "darkorange", "EC_ratio_T1" = "limegreen"),
    labels = c("part_T1" = "Participation", "GD_ratio_T1" = "GD Ratio", "EC_ratio_T1" = "EC Ratio")) +
  scale_fill_manual(
    values = c("part_T1" = "skyblue", "GD_ratio_T1" = "orange", "EC_ratio_T1" = "lightgreen"),
    labels = c("part_T1" = "Participation", "GD_ratio_T1" = "GD Ratio", "EC_ratio_T1" = "EC Ratio")) +
  scale_x_continuous(breaks = unique(coefficients$year)) +
  labs(
    x = "Années",
    y = "",
    color = "Coefficients",
    fill = "IC 95 %") +
  theme_grey() +
  theme(panel.grid.minor = element_blank())
ggsave("output_panel/ppropri_panel_coeffs.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(
    title = "Coefficients de ppropri dans les panels",
    subtitle = "(avec interactions temporelles)") +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5))  
print(plot)

rm(part_T1, GD_ratio_T1, EC_ratio_T1, coefficients, plot)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_panel_timedummies_interactions.tex"))
rm(relative_results)
```

```{r eval=FALSE}
# Pour les outcomes absolus = ratios / inscrits
time_dummies <- model.matrix(~ year * ppropri - 1, data = panel_data)
panel_as_pdata <- pdata.frame(cbind(panel_data, time_dummies), index = c("codecommune", "year"))
panel_as_pdata$year <- as.factor(panel_as_pdata$year)

regressors <- c(panel_regressors, colnames(time_dummies))
regressors <- regressors[regressors != "ppropri"]

absolute_results <- list()
for (outcome in absolute_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  absolute_results[[model_name]] <- plm(regression_formula, data = panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(absolute_results, type = "text"))
rm(model_name, regression_formula, regressors, time_dummies, outcome, panel_as_pdata)
```

```{r eval=FALSE, include=FALSE}
suppressWarnings(stargazer(absolute_results, type = "latex", out = "output_panel/absolute_panel_timedummies_interactions.tex"))
rm(absolute_results)
```

Les R2 n'évoluent pas. Pratiquement tous les sous-coefficients de ppropri sont significatifs, mais ils fluctuent d'une manière qui les rend difficilement interprétables.

## Tours confondus

### Panel simple

Comme précédemment, on obtient des R2 très faibles, bien que marginalement meilleurs.

```{r}
# Pour les outcomes relatifs = ratios / exprimés
wide_panel_as_pdata <- pdata.frame(wide_panel_data, index = c("codecommune", "year_round"))

relative_results <- list()
for (outcome in wide_relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(wide_panel_regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = wide_panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, outcome, wide_panel_as_pdata)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_widepanel_basic.tex"))
rm(relative_results)
```

### Avec dummies temporelles

```{r}
# Pour les outcomes relatifs = ratios / exprimés
wide_panel_as_pdata <- pdata.frame(wide_panel_data, index = c("codecommune", "year_round"))
wide_panel_as_pdata$year <- as.factor(wide_panel_as_pdata$year)

regressors <- c(wide_panel_regressors, "year")

relative_results <- list()
for (outcome in wide_relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = wide_panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, regressors, outcome, wide_panel_as_pdata)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_widepanel_timedummies.tex"))
rm(relative_results)
```

### Avec dummies temporelles et interactions

```{r}
# Pour les outcomes relatifs = ratios / exprimés
time_dummies <- model.matrix(~ year * ppropri - 1, data = wide_panel_data)
wide_panel_as_pdata <- pdata.frame(cbind(wide_panel_data, time_dummies), index = c("codecommune", "year_round"))
wide_panel_as_pdata$year <- as.factor(wide_panel_as_pdata$year)

regressors <- c(wide_panel_regressors, colnames(time_dummies))
regressors <- regressors[regressors != "ppropri"]

relative_results <- list()
for (outcome in wide_relative_outcomes) {
  regression_formula <- as.formula(paste(outcome, "~", paste(regressors, collapse = "+")))
  model_name <- paste0(outcome, "_model")
  relative_results[[model_name]] <- plm(regression_formula, data = wide_panel_as_pdata, model = "fd", effect = "individual", cluster = "dep", vcov = "HC1")
}

suppressWarnings(stargazer(relative_results, type = "text"))
rm(model_name, regression_formula, regressors, time_dummies, outcome, wide_panel_as_pdata)
```

```{r}
get_coeffs <- function(model, outcome_name) {
  coefficients <- as.data.frame(model$coefficients)
  coefficients <- coefficients[grep("ppropri$", rownames(coefficients)), , drop = FALSE]
  colnames(coefficients) <- "coef_ppropri"
  years <- sub("^year(\\d+):.*", "\\1", rownames(coefficients))
  coefficients$year <- years
  coefficients$outcome <- outcome_name
  
  variance <- as.data.frame(model$vcov)
  variance <- variance[grep("ppropri$", rownames(variance)), grep("ppropri$", colnames(variance)), drop = FALSE]
  variance <- diag(as.matrix(variance))
  variance <- as.data.frame(variance)
  colnames(variance) <- "var"
  
  coefficients <- cbind(coefficients, variance)
  coefficients$CI_low <- coefficients$coef_ppropri - 1.96 * sqrt(coefficients$var)
  coefficients$CI_up <- coefficients$coef_ppropri + 1.96 * sqrt(coefficients$var)
  
  return(coefficients)
}

part <- get_coeffs(relative_results$part_model, "part")
GD_ratio <- get_coeffs(relative_results$GD_ratio_model, "GD_ratio")
EC_ratio <- get_coeffs(relative_results$EC_ratio_model, "EC_ratio")
coefficients <- rbind(part, GD_ratio, EC_ratio)
coefficients$year <- as.numeric(coefficients$year)

plot <- ggplot(coefficients) +
  geom_line(aes(x = year, y = coef_ppropri, group = outcome, color = outcome)) +
  geom_line(aes(x = year, y = CI_low, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) +
  geom_line(aes(x = year, y = CI_up, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) + 
  geom_ribbon(aes(x = year, ymin = CI_low, ymax = CI_up, group = outcome, fill = outcome), alpha = 0.3) +
  geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
  scale_color_manual(
    values = c("part" = "blue", "GD_ratio" = "darkorange", "EC_ratio" = "limegreen"),
    labels = c("part" = "Participation", "GD_ratio" = "GD Ratio", "EC_ratio" = "EC Ratio")) +
  scale_fill_manual(
    values = c("part" = "skyblue", "GD_ratio" = "orange", "EC_ratio" = "lightgreen"),
    labels = c("part" = "Participation", "GD_ratio" = "GD Ratio", "EC_ratio" = "EC Ratio")) +
  scale_x_continuous(breaks = unique(coefficients$year)) +
  labs(
    x = "Années",
    y = "",
    color = "Coefficients",
    fill = "IC 95 %") +
  theme_grey() +
  theme(panel.grid.minor = element_blank())
print(plot)
ggsave("output_panel/ppropri_panels_coeffs_rounds.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot +
  labs(
    title = "Coefficients de ppropri dans les panels",
    subtitle = "(avec interactions temporelles et tours confondus)") +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5))  
print(plot)

rm(part, GD_ratio, EC_ratio, coefficients, plot)
```

```{r include=FALSE}
suppressWarnings(stargazer(relative_results, type = "latex", out = "output_panel/relative_widepanel_timedummies_interactions.tex"))
rm(relative_results)
```

# Régressions annuelles

## Simples

On peut confirmer le constat d'une variabilité des effets de ppropri en réalisant des régressions années par années, et en traçant l'évolution de son coefficient. La différence avec les panels précédents est que les coefficients des autres variables sont désormais libres au cours du temps ?

```{r}
results <- data.frame(year = integer(), "part_T1" = numeric(), "part_T2" = numeric(), "GD_ratio_T1" = numeric(), "GD_ratio_T2" = numeric(), "EC_ratio_T1" = numeric(), "EC_ratio_T2" = numeric(), "R2_adj" = numeric(), "CI_low" = numeric(), "CI_up" = numeric())

for (outcome in relative_outcomes) {
  for (yr in unique(panel_data$year)) {
    subset_data <- panel_data[panel_data$year == yr, ]
    regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
    tryCatch({
      model <- lm(regression_formula, data = subset_data)
      results <- rbind(results, data.frame(outcome = outcome, year = yr, coef_ppropri = coef(model)["ppropri"], CI_low = (confint(model)["ppropri", 1]), CI_up = (confint(model)["ppropri", 2]), r2_adj = summary(model)$adj.r.squared))
      }, error = function(e) {
        results <- rbind(results, data.frame(outcome = outcome, year = yr, coef_ppropri = NA, r2_adj = NA))
    })
  }
}

ggplot(results, aes(x = year, y = r2_adj, color = outcome, group = outcome)) +
  geom_line() +
  scale_x_continuous(breaks = unique(results$year)) +
  scale_color_manual(
    values = c("part_T1" = "blue", "GD_ratio_T1" = "darkorange", "EC_ratio_T1" = "limegreen"),
    labels = c("part_T1" = "Participation", "GD_ratio_T1" = "GD Ratio", "EC_ratio_T1" = "EC Ratio")) +
  labs(
    title = "R2 ajusté des régressions annuelles",
    x = "Années",
    y = "R2 ajusté",
    color = "Outcome") +
  theme_grey() +
  theme(
    plot.title = element_text(hjust = 0.5),
    panel.grid.minor = element_blank())

plot <- ggplot(results) +
  geom_line(aes(x = year, y = coef_ppropri, group = outcome, color = outcome)) +
  geom_line(aes(x = year, y = CI_low, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) +
  geom_line(aes(x = year, y = CI_up, group = outcome, color = outcome, ), linetype = "dashed", alpha = 0.5) + 
  geom_ribbon(aes(x = year, ymin = CI_low, ymax = CI_up, group = outcome, fill = outcome), alpha = 0.3) +
  geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
  scale_x_continuous(breaks = unique(results$year)) +
  scale_color_manual(
    values = c("part_T1" = "blue", "GD_ratio_T1" = "darkorange", "EC_ratio_T1" = "limegreen"),
    labels = c("part_T1" = "Participation", "GD_ratio_T1" = "GD Ratio", "EC_ratio_T1" = "EC Ratio")) +
  scale_fill_manual(
    values = c("part_T1" = "skyblue", "GD_ratio_T1" = "orange", "EC_ratio_T1" = "lightgreen"),
    labels = c("part_T1" = "Participation", "GD_ratio_T1" = "GD Ratio", "EC_ratio_T1" = "EC Ratio")) +
  labs(
   x = "Années",
   y = "",
   color = "Coefficient",
   fill = "IC 95 %") +
  theme_grey() +
  theme(
    plot.title = element_text(hjust = 0.5),
    panel.grid.minor = element_blank())
ggsave("output_panel/ppropri_reg_coeffs.png", plot, width = 8, height = 6, dpi = 300)

plot <- plot + labs(title = "Coefficients de ppropri dans les régressions annuelles")
print(plot)

rm(results, outcome, yr, subset_data, regression_formula, model, plot)
```

## Pseudo-quantiles

En considérant les quantiles de ppropri plutôt que de l'outcome.

```{r}
# Possibilité 1 : considérer la distribution relative avec les quantiles années par année
pseudo_quant_reg <- function(outcome) {
  coefs_df <- data.frame(year = integer(), quantile = character(), coefficient = numeric())
  R2_df <- data.frame(year = integer(), quantile = character(), R2 = numeric())
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
  
  for (yr in unique(panel_data$year)) {
    panel_subset <- panel_data[panel_data$year == yr, ]
    quantiles <- quantile(panel_subset$ppropri, probs = seq(0, 1, by = 0.25), na.rm = TRUE)
    for (i in 1:(length(quantiles) - 1)) {
      quantile_subset <- panel_subset[panel_subset$ppropri >= quantiles[i] & panel_subset$ppropri <= quantiles[i + 1], ]
      model <- lm(regression_formula, data = quantile_subset)
      coefs_df <- rbind(coefs_df, data.frame(year = yr, quantile = paste(i), coefficient = model$coef["ppropri"]))
      R2_df <- rbind(R2_df, data.frame(year = yr, quantile = paste(i), R2 = summary(model)$adj.r.squared))
    }
  }
  
  plot_r2 <- ggplot(R2_df, aes(x = year, y = R2, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(R2_df$year)) + 
    labs(
      title = paste("R2 des régressions pseudo-quantiles de", outcome),
      subtitle ="(quantiles mobiles)",
      x = "Années",
      y = "R2",
      color = "Quartiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  print(plot_r2)
  
  output_file <- paste0("output_panel/pseudo_quant_mob_", outcome, ".png")
  plot_coef <- ggplot(coefs_df, aes(x = year, y = coefficient, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(coefs_df$year)) +
    labs(
      x = "Années",
      y = "Coefficients",
      color = "Quartiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  ggsave(output_file, plot_coef, width = 8, height = 6, dpi = 300)
  
  plot_coef <- plot_coef +
    labs(
      title = paste("Coefficients de ppropri dans les régressions pseudo-quantiles de", outcome),
      subtitle = "(quantiles mobiles)")
  print(plot_coef)
}

for (outcome in relative_outcomes) {pseudo_quant_reg(outcome)}

rm(pseudo_quant_reg, outcome)
```

```{r}
# Possibilité 2 : considérer les niveaux absolus
pseudo_quant_reg <- function(outcome) {
  coefs_df <- data.frame(year = integer(), quantile = character(), coefficient = numeric())
  R2_df <- data.frame(year = integer(), quantile = character(), R2 = numeric())
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
  
  ppropri_ranges <- list("< 60 %" = c(0, 60),
                         "60-80 %" = c(60, 80),
                         "80-95 %" = c(80, 95),
                         "95 % <" = c(95, Inf))
  
  for (yr in unique(panel_data$year)) {
    panel_subset <- panel_data[panel_data$year == yr, ]
    for (range_name in names(ppropri_ranges)) {
      range_values <- ppropri_ranges[[range_name]]
      subset_name <- paste(range_name)
      quantile_subset <- panel_subset[panel_subset$ppropri >= range_values[1] & panel_subset$ppropri < range_values[2], ]
      if (nrow(quantile_subset) > 0) {
        model <- lm(regression_formula, data = quantile_subset)
        coefs_df <- rbind(coefs_df, data.frame(year = yr, quantile = subset_name, coefficient = model$coef["ppropri"]))
        R2_df <- rbind(R2_df, data.frame(year = yr, quantile = subset_name, R2 = summary(model)$adj.r.squared))
      }
    }
  }
  
  plot_r2 <- ggplot(R2_df, aes(x = year, y = R2, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(R2_df$year)) +  
    labs(
      title = paste("R2 des régressions pseudo-quantiles de", outcome),
      subtitle = "(quantiles fixes)",
      x = "Années",
      y = "R2",
      color = "Quantiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  print(plot_r2)
  
  output_file <- paste0("output_panel/pseudo_quant_fixed_", outcome, ".png")
  plot_coef <- ggplot(coefs_df, aes(x = year, y = coefficient, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(coefs_df$year)) + 
    labs(
      x = "Années",
      y = "Coefficients",
      color = "Quantiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  ggsave(output_file, plot_coef, width = 8, height = 6, dpi = 300)
  
  plot_coef <- plot_coef +
    labs(
      title = paste("Coefficients de ppropri dans les régressions pseudo-quantiles de", outcome),
      subtitle = "(quantiles fixes)")
  print(plot_coef)
}

for (outcome in relative_outcomes) {pseudo_quant_reg(outcome)}

rm(pseudo_quant_reg, outcome)
```

```{r}
ppropri_df <- data.frame(year = integer(), sum_comm = numeric(), sum_pop = numeric())

for (yr in unique(panel_data$year)) {
  panel_subset <- panel_data[panel_data$year == yr, ]
  panel_subset <- panel_subset[panel_subset$ppropri >= 95, ]
  sum_comm <- nrow(panel_subset)
  sum_pop <- sum(panel_subset$popcomm, na.rm = TRUE)
  ppropri_df <- rbind(ppropri_df, data.frame(year = yr, sum_comm = sum_comm, sum_pop = sum_pop))
}

plot_counts <- ggplot(ppropri_df, aes(x = year, y = sum_comm)) +
  geom_line(color = "skyblue") +
  scale_x_continuous(breaks = unique(ppropri_df$year)) +  
  labs(
    x = "Années",
    y = "Effectifs") +
  theme_grey() +
  theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
ggsave("output_panel/ppropri_max.png", plot_counts, width = 8, height = 6, dpi = 300)

plot_counts <- plot_counts + labs(title = "Nombre de communes où le taux de propriétaires est >95%")
print(plot_counts)

rm(ppropri_df, plot_counts, yr, sum_comm, sum_pop)
```

## Régressions quantiles

```{r}
# Possibilité 1 : considérer la distribution relative avec les quantiles années par année
quant_reg <- function(outcome) {
  coefs_df <- data.frame(year = integer(), coefficient = numeric(), quantile = numeric())
  R2_df <- data.frame(year = integer(), R2 = numeric(), quantile = numeric())
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))

  quantiles <- c(0.2, 0.4, 0.6, 0.8)
    
  for (yr in unique(panel_data$year)) {
    subset_data <- panel_data[panel_data$year == yr, ]
    for (q in quantiles) {
      model <- rq(regression_formula, data = subset_data, tau = q)
      coefs_df <- rbind(coefs_df, data.frame(year = yr, coefficient = coef(model)["ppropri"], quantile = q))
    }
  }
  
  output_file <- paste0("output_panel/quant_mob_", outcome, ".png")
  plot_coef <- ggplot(coefs_df, aes(x = year, y = coefficient, color = factor(quantile))) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(coefs_df$year)) +   
    labs(
      x = "Années",
      y = "Coefficients") +
    scale_color_discrete(name = "Quantile") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  ggsave(output_file, plot_coef, width = 8, height = 6, dpi = 300)
  
  plot_coef <- plot_coef +
    labs(
      title = paste("Coefficients de ppropri dans les régressions quantiles de", outcome),
      subtitle = "(quantiles mobiles)")
  print(plot_coef)
}

for (outcome in relative_outcomes) {quant_reg(outcome)}

rm(quant_reg, outcome)
```

```{r}
# Possibilité 2 : considérer les niveaux absolus
quant_reg <- function(outcome) {
  coefs_df <- data.frame(year = integer(), quantile = character(), coefficient = numeric())
  R2_df <- data.frame(year = integer(), quantile = character(), R2 = numeric())
  regression_formula <- as.formula(paste(outcome, "~", paste(panel_regressors, collapse = "+")))
  
  outcome_ranges <- list("< 10 %" = c(0, 10),
                         "10-20 %" = c(10, 20),
                         "20-30 %" = c(20, 30),
                         "30-40 %" = c(30, 40),
                         "40-50 %" = c(40, 50),
                         "50-60 %" = c(50, 60),
                         "60-70 %" = c(60, 70),
                         "70-80 %" = c(70, 80),
                         "80-00 %" = c(80, 90),
                         "90 % <" = c(90, Inf))
  
  for (yr in unique(panel_data$year)) {
    panel_subset <- panel_data[panel_data$year == yr, ]
    for (range_name in names(outcome_ranges)) {
      range_values <- outcome_ranges[[range_name]]
      quantile_subset <- panel_subset[panel_subset[[outcome]] >= range_values[1] & panel_subset[[outcome]] < range_values[2], ]
      tryCatch({
        model <- lm(regression_formula, data = quantile_subset)
        coefs_df <- rbind(coefs_df, data.frame(year = yr, quantile = range_name, coefficient = model$coef["ppropri"]))
        R2_df <- rbind(R2_df, data.frame(year = yr, quantile = range_name, R2 = summary(model)$adj.r.squared))
      }, error = function(e) {
        coefs_df <- rbind(coefs_df, data.frame(year = yr, quantile = range_name, coefficient = NA))
        R2_df <- rbind(R2_df, data.frame(year = yr, quantile = range_name, R2 = NA))
      })
    }
  }
  
  plot_r2 <- ggplot(R2_df, aes(x = year, y = R2, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(R2_df$year)) +  
    labs(
      title = paste("R2 des régressions pseudo-quantiles de", outcome),
      subtitle ="(quantiles fixes)",
      x = "Années",
      y = "R2",
      color = "Quantiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  print(plot_r2)
  
  output_file <- paste0("output_panel/quant_fixed_", outcome, ".png")
  plot_coef <- ggplot(coefs_df, aes(x = year, y = coefficient, color = quantile, group = quantile)) +
    geom_line() +
    geom_hline(yintercept = 0, linetype = "twodash", color = "gray40") + 
    scale_x_continuous(breaks = unique(coefs_df$year)) + 
    labs(
      x = "Années",
      y = "Coefficients",
      color = "Quantiles") +
    theme_grey() +
    theme(
      plot.title = element_text(hjust = 0.5),
      plot.subtitle = element_text(hjust = 0.5),
      legend.title = element_text(hjust = 0.5),
      panel.grid.minor = element_blank())
  ggsave(output_file, plot_coef, width = 8, height = 6, dpi = 300)
  
  plot_coef <- plot_coef +
    labs(
      title = paste("Coefficients de ppropri dans les régressions pseudo-quantiles de", outcome),
      subtitle ="(quantiles fixes)")
  print(plot_coef)
}

for (outcome in relative_outcomes) {quant_reg(outcome)}

rm(quant_reg, outcome)
```
